requirements.txt
aiofiles==22.1.0
anyio==3.6.2
argcomplete==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
backports.cached-property==1.0.2
beautifulsoup4==4.11.1
bleach==5.0.1
boto3==1.25.4
botocore==1.28.4
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
clip==1.0
contourpy==1.0.5
cycler==0.11.0
dacite==1.6.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.6
dnspython==2.2.1
entrypoints==0.4
eventlet==0.33.1
executing==1.2.0
fastjsonschema==2.16.2
fiftyone==0.17.2
fiftyone-brain==0.9.1
fiftyone-db==0.3.0
filelock==3.8.0
fonttools==4.38.0
ftfy==6.1.1
future==0.18.2
glob2==0.7
graphql-core==3.1.7
greenlet==1.1.3.post0
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.15.0
httpx==0.23.0
huggingface-hub==0.10.1
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
imageio==2.22.2
importlib-metadata==5.0.0
ipykernel==6.16.2
ipython==8.6.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.10
jsonschema==4.16.0
jupyter-server==1.21.0
jupyter_client==7.4.4
jupyter_core==4.11.2
jupyterlab==3.5.0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.1
kaleido==0.2.1
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
mistune==2.0.4
mongoengine==0.20.0
motor==2.5.1
nbclassic==0.4.7
nbclient==0.7.0
nbconvert==7.2.3
nbformat==5.7.0
ndjson==0.3.1
nest-asyncio==1.5.6
networkx==2.8.7
notebook==6.5.2
notebook_shim==0.2.0
numpy==1.23.4
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
patool==1.12
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
plotly==5.11.0
pprintpp==0.4.0
priority==2.0.0
prometheus-client==0.15.0
prompt-toolkit==3.0.31
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
pymongo==3.12.3
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
regex==2022.9.13
requests==2.28.1
retrying==1.3.3
rfc3986==1.5.0
s3transfer==0.6.0
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
Send2Trash==1.8.0
sentinel==0.3.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sse-starlette==0.10.3
sseclient-py==1.7.2
stack-data==0.6.0
starlette==0.16.0
strawberry-graphql==0.96.0
tabulate==0.9.0
tenacity==8.1.0
terminado==0.17.0
threadpoolctl==3.1.0
tifffile==2022.10.10
tinycss2==1.2.1
tokenizers==0.13.1
toml==0.10.2
tomli==2.0.1
torch==1.13.0+cu116
torchaudio==0.13.0+cu116
torchvision==0.14.0+cu116
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
transformers==4.23.1
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
universal-analytics-python3==1.1.1
urllib3==1.26.12
voxel51-eta==0.8.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
wrapt==1.14.1
wsproto==1.2.0
xmltodict==0.13.0
zipp==3.10.0
